refactor: 拆分 polynomial_factorize.hh 为 4 个文件（方案 C）
将 2898 行的单体头文件按职责拆分：
- polynomial_factorize_zp.hh    (395行): Zp 辅助 + M1 squarefree/DDF/EDF
- polynomial_factorize_univar.hh(958行): factorization<> + ZZ 辅助 + M2 Hensel
                                         + M3 Zassenhaus + 素数选择 + factorize(upolynomial)
- polynomial_factorize_wang.hh  (1369行): taylor_coeff + Wang LC + 多变量 Hensel
                                          + __wang_core + __factor_multivar
- polynomial_factorize.hh       (235行): 三个 #include + 公开 factorize() API

额外清理：删除两个不再需要的前向声明
(__symmetric_mod 和 __factor_multivar)

docs: 添加 van Hoeij LLL 调研报告 (P1 调研阶段)
